+++ tests/test_add_error.py
@@ -108,3 +108,25 @@ def test_add_error_log_traceback_variables(suite, suite_test, log_variables, con
         assert found == log_variables, 'Variable {!r} not found in traceback log!'.format(variable_name)
     _search_variable('x_variable', 'xxx')
     _search_variable('self.property_value', 'yyy')
+
+
+def test_add_error_log_traceback_variables_self_none(suite, suite_test, config_override, tmpdir):
+    config_override('log.core_log_level', logbook.TRACE)
+    config_override('log.traceback_variables', True)
+    config_override('log.root', str(tmpdir.join('logs')))
+
+    @suite_test.prepend_body
+    def __code__():          # pylint: disable=unused-variable
+        # to avoid the line itself from being detected
+        self = None # pylint: disable=unused-variable
+
+    suite_test.when_run.error()
+    res = suite.run()
+    result = res[suite_test]
+
+    with open(result.get_log_path()) as f:
+        lines = f.read()
+
+    assert 'self: None' in lines
+    [err] = result.get_errors()
+    assert 'Test exception' in str(err)