+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :bug:`767` Fixed traceback variable capture for cases where ``self=None``
 * :release:`1.5.0 <7-3-2018>`
 * :feature:`590` Add support for labeling parametrization variations
 * :feature:`697` Added ``slash.before_interactive_shell`` hook